move date-related functions and activity helpers to better-fitting places
