+++ client/src/components/TodayTimeline/useTimeline.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
+import { currentFractionOfDay } from "../../lib/datetime/day-fraction";
 import { defaultWidth } from "./mock";
-import { currentFractionOfDay } from "./now";
 
 function useTimeline() {
 	const [width, setWidth] = useState<number>(defaultWidth); // pixel width of timeline body
