+++ client/src/lib/activity.ts
@@ -1,8 +1,8 @@
 import dayjs from "dayjs";
-import { Datelike } from "../../types/date.types";
-import { ActivityWithIds } from "../../types/server/activity.types";
-import { ID } from "../../types/server/utility.types";
-import { sameDay } from "./date";
+import { Datelike } from "../types/date.types";
+import { ActivityWithIds } from "../types/server/activity.types";
+import { ID } from "../types/server/utility.types";
+import { sameDay } from "./datetime/compare";
 
 /** Gets the `start` of an activity, which is either a timestamp or
  * year-month-date string. */
