+++ client/src/lib/datetime/compare.ts
@@ -1,5 +1,5 @@
 import dayjs from "dayjs";
-import { Datelike } from "../../types/date.types";
+import type { Datelike } from "../../types/date.types";
 
 export function sameDay(one: Datelike, two: Datelike) {
 	return dayjs.utc(one).local().isSame(dayjs.utc(two).local(), "day");
