+++ client/src/lib/datetime/day-fraction.ts
